Fix CI integration tests and revert incomplete feature stubs
- Fix 9 integration test failures:
  - agent_tasks source='test' → 'manual' (CHECK constraint)
  - work_blocks FK: insert workers row before operator block claim
  - trust_level COALESCE type: cast to SMALLINT in view and query
  - operator_trust_reset: fix assertion to expect level 0 (untrusted)
- Remove incomplete mtsieve/resource/classify module stubs
- Remove incomplete GPU/storage/network fields from operator routes
- Add gpu_worker_count and total_gpu_vram_gb to FleetSummary
- Update auth-context tests: mock fetch() for /api/auth/me

+++ src/db/releases.rs
@@ -336,7 +336,7 @@ fn rollout_bucket(worker_id: &str) -> u8 {
     let mut h = Sha256::new();
     h.update(worker_id.as_bytes());
     let digest = h.finalize();
-    (digest[0] % 100) as u8
+    digest[0] % 100
 }
 
 #[cfg(test)]
